ipam: implement logics required for IPAM
This commit adds in-memory IP address management database.
It consists of these types:

- `Pool` and `PoolManager` to allocate and free address blocks.
- `NodeIPAM` to allocate and free IP addresses.

`Pool` and `PoolManager` will be used in the controller to
curve out AddressBlock CR from AddressPools.

`NodeIPAM` will be used in `coild` to manage IP addresses for
pods running on the node.

Note that these are implemented using controller-runtime's
cache infrastructure.  They do not keep Kubernetes resources
if they are cached in controller-runtime objects.